requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
Flask==0.10.1
Flask-Cache==0.13.1
itsdangerous==0.24
Jinja2==2.8
jsonpickle==0.9.2
kombu==3.0.26
MarkupSafe==0.23
PyGithub==1.25.2
pytz==2015.4
PyYAML==3.11
redis==2.10.3
semantic-version==2.4.2
uWSGI==2.0.11.1
Werkzeug==0.10.4
wheel==0.24.0
